Move config normalization methods to a module that is not extended by jwplayer-commercial
