+++ src/js/api/config.js
@@ -1,3 +1,4 @@
+import { normalizeAspectRatio, normalizeSize } from 'api/config-normalization';
 import { loadFrom, getScriptPath } from 'utils/playerutils';
 import { serialize } from 'utils/parser';
 import { isValidNumber, isNumber, pick, isBoolean } from 'utils/underscore';
@@ -44,13 +45,6 @@ function _deserialize(options) {
         }
         options[key] = serialize(options[key]);
     });
-}
-
-export function normalizeSize(val) {
-    if (val.slice && val.slice(-2) === 'px') {
-        val = val.slice(0, -2);
-    }
-    return val;
 }
 
 
@@ -168,27 +162,5 @@ const Config = function(options, persisted) {
     config.backgroundLoading = isBoolean(config.backgroundLoading) ? config.backgroundLoading : Features.backgroundLoading;
     return config;
 };
-
-export function normalizeAspectRatio(ar, width) {
-    if (width.toString().indexOf('%') === -1) {
-        return 0;
-    }
-    if (typeof ar !== 'string' || !ar) {
-        return 0;
-    }
-    if (/^\d*\.?\d+%$/.test(ar)) {
-        return ar;
-    }
-    const index = ar.indexOf(':');
-    if (index === -1) {
-        return 0;
-    }
-    const w = parseFloat(ar.substr(0, index));
-    const h = parseFloat(ar.substr(index + 1));
-    if (w <= 0 || h <= 0) {
-        return 0;
-    }
-    return (h / w * 100) + '%';
-}
 
 export default Config;
